move error types inside participant
